+++ Octree/Octree.h
@@ -46,6 +46,6 @@ public:
 	void			DrawAllBoxes(CShader* myShader);
 	void			DrawOctreeLeaves(CShader* myShader);
 	bool			isColliding(glm::vec3 point);
-	bool			isColliding(double radius, glm::vec3 centerPoint, glm::vec3 AABBCenter);
+	bool			isColliding(float radius, glm::vec3 centerPoint, glm::vec3 AABBCenter);
 
 };
